refactor(pool): redesign emulation and pool ID strategy
