+++ src/client/conn/tls_info.rs
@@ -4,7 +4,7 @@ use tokio::net::TcpStream;
 use tokio::net::UnixStream;
 use tokio_btls::SslStream;
 
-use crate::tls::{MaybeHttpsStream, TlsInfo};
+use crate::tls::{TlsInfo, conn::MaybeHttpsStream};
 
 /// A trait for extracting TLS information from a connection.
 ///
