+++ src/client/layer.rs
@@ -1,5 +1,6 @@
 //! Middleware for the client.
 
+pub mod client;
 pub mod config;
 #[cfg(feature = "cookies")]
 pub mod cookie;
